fix unit tests

+++ resource/generation/testdata/migrations/000001_DataChangeEvents.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE DataChangeEvents (
   TableName STRING(MAX) NOT NULL,
   RowId STRING(MAX) NOT NULL,
-  Sequence INTEGER NOT NULL,
+  Sequence INT64 NOT NULL,
   EventTime TIMESTAMP NOT NULL OPTIONS (allow_commit_timestamp=true),
   EventSource STRING(MAX) NOT NULL,
   ChangeSet JSON,
